+++ include/dxfeed_graal_cpp_api/event/candle/CandlePrice.hpp
@@ -127,7 +127,7 @@ struct DXFCPP_EXPORT CandlePrice : CandleSymbolAttribute {
      * @param symbol The candle symbol string.
      * @return candle price type of the given candle symbol string.
      */
-    static std::reference_wrapper<const CandlePrice> getAttributeForSymbol(const StringLike &symbol) noexcept;
+    static std::reference_wrapper<const CandlePrice> getAttributeForSymbol(const StringLike &symbol);
 
     /**
      * Returns candle symbol string with the normalized representation of the candle price type attribute.